Multi tenant - access rights

+++ app/api/storage_endpoint.py
@@ -1,9 +1,10 @@
 from typing import Optional
 
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 
 from app.api.auth.permissions import Permissions
 from app.config import server
+from tracardi.config import tracardi
 from tracardi.service.storage.driver import storage
 from tracardi.service.storage.elastic_client import ElasticClient
 from tracardi.service.storage.indices_manager import check_indices_mappings_consistency
@@ -36,6 +37,10 @@ async def get_index_mapping_metadata(index: str, filter: str = None):
     """
     Returns metadata of given index (str)
     """
+
+    if tracardi.multi_tenant:
+        raise HTTPException(status_code=405, detail="This operation is not allowed for multi-tenant server.")
+
     result = await storage.driver.raw.get_mapping_fields(index)
     if filter is not None:
         result = [item for item in result if item.startswith(filter) and item!=filter]
@@ -47,6 +52,10 @@ async def get_index_mapping(index: str):
     """
     Returns mapping of given index (str)
     """
+
+    if tracardi.multi_tenant:
+        raise HTTPException(status_code=405, detail="This operation is not allowed for multi-tenant server.")
+
     mapping = await storage.driver.raw.get_mapping(index)
     return mapping.get_field_names()
 
@@ -66,6 +75,9 @@ async def reindex_data(source: str, destination: str, wait_for_completion: bool 
     Copies data from one index to another.
     """
 
+    if tracardi.multi_tenant:
+        raise HTTPException(status_code=405, detail="This operation is not allowed for multi-tenant server.")
+
     return await storage.driver.raw.reindex(source, destination, wait_for_completion)
 
 
@@ -74,6 +86,9 @@ async def delete_index(index_name: str):
     """
     Deletes storage index
     """
+
+    if tracardi.multi_tenant:
+        raise HTTPException(status_code=405, detail="This operation is not allowed for multi-tenant server.")
 
     es = ElasticClient.instance()
     return await es.remove_index(index_name)
@@ -89,6 +104,9 @@ async def get_snapshot_repository(name: str):
     List repository snapshots
     """
 
+    if tracardi.multi_tenant:
+        raise HTTPException(status_code=405, detail="This operation is not allowed for multi-tenant server.")
+
     return await storage.driver.snapshot.get_snapshot_repository(repo=name)
 
 
@@ -98,5 +116,8 @@ async def get_snapshot_repository_status(name: Optional[str] = "_all"):
     """
     Lists available snapshots withing the repository name. Use _all as a name to get all repos snapshots.
     """
+
+    if tracardi.multi_tenant:
+        raise HTTPException(status_code=405, detail="This operation is not allowed for multi-tenant server.")
 
     return await storage.driver.snapshot.get_repository_snapshots(repo=name)
